# TaskManagement.DTO/Office/User/UserUpdateDto.cs
namespace TaskManagement.DTO.Office.User;

public record UserUpdateDto {
    public Guid Id { get; init; }
    
    public string? FirstName { get; init; }
    public string? LastName { get; init; }
    
    public string? Email { get; init; }
    public string? Password { get; init; }
}